requirements.txt
easydict
st_moe_pytorch
timm
tqdm==4.65.2
pandas
opencv-python
scikit-learn